Remove the Connection from SqlStament, moving it the execution logic to new StatementExecutor classes
